scripts/requirements.txt
# Requirements for OAB Dataset Import Scripts
# =============================================

# Core dependencies
pandas
datasets
supabase

# Data processing
numpy
python-dotenv

# HTTP and API clients
httpx
requests

# JSON and data handling
pydantic

# Progress tracking
tqdm

# Logging
loguru

# Optional: for advanced data analysis
scikit-learn
matplotlib
seaborn
